bullet are now drawn rotated

+++ src/main/java/fr/poulpogaz/jam/patterns/LinearPattern.java
@@ -11,7 +11,7 @@ public class LinearPattern implements MovePattern {
     }
 
     @Override
-    public Vector2f dir(float t) {
+    public Vector2f dir(int t) {
         return dir;
     }
 }
