+++ src/main/java/fr/poulpogaz/jam/patterns/StaticPattern.java
@@ -4,7 +4,7 @@ import org.joml.Vector2f;
 
 public class StaticPattern implements MovePattern {
     @Override
-    public Vector2f dir(float t) {
+    public Vector2f dir(int t) {
         return new Vector2f();
     }
 }
